% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add 2 numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{param 1}

\item{y}{param 2}
}
\value{
sum of 2 numbers
}
\description{
description demo
}
\details{
details demo
}
\examples{
add(1,2)
\donttest{
add(3,4)
}
}
