% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDregular.R
\name{BDregular}
\alias{BDregular}
\title{Accuracy matrix block diagonal regularization estimation}
\usage{
BDregular(X, dim = nrow(X), lam = "ada", shrink = "MCP", min.eig = 0.001)
}
\arguments{
\item{x}{param 1}
}
\value{
Precision matrix estimate
}
\description{
description demo
}
\details{
details demo
}
\examples{
BDregular(X)
}
